+++ src/inference.ts
@@ -1,12 +1,28 @@
 import OpenAI from 'openai';
 import 'dotenv/config';
 import type { Config } from './config';
-import type { Message } from './types';
 
-const systemMessage = {
+export type Message = SystemMessage | UserMessage | AiMessage;
+
+export interface SystemMessage {
+  role: 'system';
+  content: string;
+}
+
+export interface UserMessage {
+  role: 'user';
+  content: string;
+}
+
+export interface AiMessage {
+  role: 'assistant';
+  content: string;
+}
+
+const systemMessage: SystemMessage = {
   role: 'system',
   content: 'You are a helpful assistant responding in a concise manner to user questions.',
-} as const;
+};
 
 export async function getChatCompletion(config: Config, messages: Message[]) {
   const openai = new OpenAI({
