--- src/types.ts
@@ -1,16 +1,0 @@
-export type Message = SystemMessage | UserMessage | AiMessage;
-
-export interface SystemMessage {
-  role: 'system';
-  content: string;
-}
-
-export interface UserMessage {
-  role: 'user';
-  content: string;
-}
-
-export interface AiMessage {
-  role: 'assistant';
-  content: string;
-}
